Remove pipeline dir
This didn't really do much, all the methods can
easily just be in the main.

io is small enough and important enough to just be
to-level in my eyes.

- Let test code replace CRLF with LF line endings,
  so we don't constantly have to deal with this.
